requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
groq
httpx
duckduckgo-search
jinja2